requirement.txt
Package                Version
---------------------- ------------
asgiref                3.7.2
Django                 5.0.1
django-cors-headers    4.3.1
django-rest-framework  0.1.0
djangorestframework    3.14.0
future                 0.18.3
mysql-connector-python 8.2.0
mysqlclient            2.2.1
odict                  1.9.0
pillow                 10.2.0
pip                    24.0
plumber                1.7
protobuf               4.21.12
PyMySQL                1.1.0
pytz                   2023.3.post1
setuptools             69.0.3
sqlparse               0.4.4
tzdata                 2023.4
zope.component         6.0
zope.deferredimport    5.0
zope.deprecation       5.0
zope.event             5.0
zope.hookable          6.0
zope.interface         6.1
zope.lifecycleevent    5.0
zope.proxy             5.1
